rename:AlbumSearch 컴포넌트 관련 디렉토리생성
* /components/album/AlbumSearch
* AlbumSearch, AlbumSearcResultList 컴포넌트 이동
