+++ src/pages/MainPage.jsx
@@ -7,7 +7,7 @@ import ReviewMain from '../components/review/ReviewMain';
 import FloatingActionButton from '../components/common/FloatingActionButton';
 import { useAuth } from '../AuthenticationContext';
 import EditProfile from '../components/common/EditProfile';
-import AlbumSearch from '../components/album/AlbumSearch';
+import AlbumSearch from '../components/album/AlbumSearch/AlbumSearch';
 import Banner from '../components/common/Banner/Banner';
 
 function MainPage() {
